% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{add.dates}
\alias{add.dates}
\title{Adding sampling years based on date of visit and panel}
\usage{
add.dates(pts)
}
\arguments{
\item{pts}{The SPDF to add YEAR to.}
}
\description{
For the given SPDF, this adds the year extracted from the values in PANEL, then overwrites that with the year extracted from DT_VST if possible, then for all remaining observations still missing a YEAR value it makes a best guess based on the YEAR values in the other observations with the same PANEL value.
}
