% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{add_area}
\alias{add_area}
\title{Add areas to a polygon or multipolygon sf object}
\usage{
add_area(polygons, area_ha = TRUE, area_sqkm = TRUE, byid = TRUE)
}
\arguments{
\item{polygons}{Polygon or multipolygon sf object. These are the polygons to add the area variables to. Areas will be added to each observation in the attributes, so make sure that polygons are "dissolved" by their identities before calculating areas.}

\item{area_ha}{Logical. If \code{TRUE}, areas will be calculated and added in hectares. Default is \code{TRUE}.}

\item{area_sqkm}{Logical. If \code{TRUE}, areas will be calculated and added in square kilometers. Default is \code{TRUE}.}
}
\value{
The original sf object with an additional field for each area unit calculated, always with at least \code{"area_sqm"}.
}
\description{
This function takes a polygon or multipolygon sf object and calculates and adds area fields to the data frame. It will always add square meters, but may additionally add square kilometers and hectares as well.
}
\keyword{area}
