% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzing.R
\name{analyze_cat}
\alias{analyze_cat}
\title{Estimation of weighted proportions of categorical data}
\usage{
analyze_cat(
  data,
  weights,
  id_var,
  cat_var,
  wgt_var,
  definitions = NULL,
  conf = 80,
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame. Categorical data with the unique identifiers for each observation/row in the variable \code{id_var} and the assigned category for each observation/row in \code{cat_var}. Note that the unique identifiers are the link between \code{data} and \code{weights}}

\item{weights}{Data frame. This must contain the weighting information using the variables \code{id_var} with a unique identifier for each observation/row and \code{wgt_var} with the relative numeric weight of each observation/row.}

\item{id_var}{Character string. The name of the variable in \code{data} and \code{weights} that contains the unique identifiers for the observations. All values in \code{data$id_var} must appear in \code{weights$id_var}.}

\item{cat_var}{Character string. The name of the variable in \code{data} that contains the category values as character strings.}

\item{wgt_var}{Character string. The name of the variable in \code{weights} that contains the numeric weight values.}

\item{definitions}{Conditionally optional character vector. The possible categories that the observation could've been classed into. This is NOT optional if there are categories that do not appear in \code{data} because no observations met their criteria because those categories must be included in the calculations. Must contain at least the values in \code{code$cat_var} but should include ALL possible categories.}

\item{conf}{Numeric. The confidence level in percent. Defaults to \code{80}.}

\item{verbose}{Logical. If \code{TRUE} then the function will generate additional messages as it executes. Defaults to \code{FALSE}.}
}
\value{
A data frame containing the categories, counts of observations, weighted estimated proportions, and confidence intervals.
}
\description{
Given categorical data and the weights for the individual observations, calculate estimated proportions by category and Goodman's multinomial confidence intervals.
}
