% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{analyzer}
\alias{analyzer}
\title{Categorical Analysis of AIM Data}
\usage{
analyzer(evaluated.points, weights,
  default.reportingunit = "No reporting unit", reportingunit.type = NA,
  adjustedweights = T)
}
\arguments{
\item{evaluated.points}{Data frame output from benchmarker()}

\item{weights}{Data frame output from weighter(), equivalent to weighter()[["point.weights"]] or weighter()[[2]]. Optionally, the output from weight.adjuster() instead}

\item{adjustedweights}{Logical, defaults to TRUE. If the field ADJWGT is found in the data frame provided for the argument weights, should it be used instead of the field WGT? Allows for quick comparison between the results of analysis with and without the weight adjustment}

\item{tdat}{Data frame or SpatialPointsDataFrame containing the TerrADat PrimaryKey field and reporting units for the points found in evaluated.points}
}
\value{
A data frame with the
}
\description{
This function takes the outputs from the functions weighter() optionally passed through weight.adjuster() benchmarker(), and a TerrADat SpatialPointsDataFrame with a field for reporting units. It applies spsurvey::cat.analysis() and returns the results
}
\examples{
analyzer()
}
\keyword{analysis}
