% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{apply.tracking}
\alias{apply.tracking}
\title{Apply project tracking Excel files to imported Design Databases.}
\usage{
apply.tracking(filenames, path = "", dd.list, dd.names = c(""), tdat,
  target.values = c("Target Sampled", "TS"), deleteoverdraw = T)
}
\arguments{
\item{filenames}{A character vector of the filenames (including extension) of the project tracking Excel files to import. If not using the \code{path} argument, the filename should include the entire filepath.}

\item{path}{Optional string specifying a common filepath containing the project tracking sheets to read in. This will be prepended to the values in \code{filenames}. If the tracking sheets are in different folder paths, do not provide this.}

\item{dd.list}{Output from \code{read.dd()}.}

\item{dd.names}{An optional character string vector of Design Database names from \code{dd.list} to compare against the plot tracking Excel files. If not provided, all of the Design Databases represented in \code{dd.list} will be compared and updated.}

\item{tdat}{Output from \code{read.tdat()}.}

\item{target.values}{Character string or character vector. This defines what values in the point fate field count as target points. The function always looks for "Target Sampled" and "TS", so this argument is only necessary if there are additional values in the sample design databases. This is case insensitive}

\item{deleteoverdraw}{Logical. If \code{T} then unsampled overdraw points will be dropped. Defaults to \code{T}.}
}
\value{
Returns \code{dd.list} with sampling dates and statuses added from the plot tracking Excel files.
}
\description{
Imports plot tracking worksheets used during an AIM project and uses them to assigns statuses to imported Design Databases.
}
