% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{apply_benchmarks}
\alias{apply_benchmarks}
\title{Apply benchmarks to data}
\usage{
apply_benchmarks(
  data,
  benchmarks,
  data_id_vars,
  data_group_var = NULL,
  benchmark_group_var = "Benchmark_Group",
  benchmark_indicator_var = "Indicator",
  use_tdat_indicator_lookup = TRUE
)
}
\arguments{
\item{data}{Data frame. The data to be benchmarked. Must contain the variable(s) in \code{data_id_vars} as unique identifiers. May be "wide" or "long" data. If wide, then each indicator name should be a variable name. If long, then the indicator names and values should be in variables named \code{"indicator_var"} and \code{"value"}.}

\item{benchmarks}{Data frame. The benchmarks in the format produced by \code{read_benchmarks()}.}

\item{data_id_vars}{Character string or vecor of character strings. The names of the variable(s) in \code{data} which serve as unique identifiers.}

\item{data_group_var}{Optional character string. The name of the variable that holds the benchmark group memberships. If \code{NULL} then it defaults to the variable in \code{benchmark_group_var}. Defaults to \code{NULL}.}

\item{benchmark_group_var}{Character string. The name of the variable that holds the benchmark group memberships. Defaults to \code{"Benchmark_Group"}.}

\item{benchmark_indicator_var}{Character string. The name of the variable that holds the indicator names in the benchmarks. Defaults to \code{"Indicator"}.}

\item{use_tdat_indicator_lookup}{Logical. Set to \code{TRUE} if the the values in variable \code{benchmark_indicator_var} are the standard human-readable names for the indicators and the corresponding variables in \code{data} use the variable names from the Terrestrial AIM Database. Defaults to \code{TRUE}.}
}
\value{
A long data frame of each point/benchmarked indicator combination with the assigned condition category.
}
\description{
Using data that has unique identifier and indicator value variables, apply benchmarks. The benchmarks are applied to the data based on the benchmark group(s) that the data belong to, which may already be assigned in the data or may be added with a benchmark group lookup table.
}
