% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute.shapefile.R
\name{attribute.shapefile}
\alias{attribute.shapefile}
\title{Attributing a Spatial Data Frame using another Spatial Data Frame}
\usage{
attribute.shapefile(shape1, shape2, attributefield = NULL, newfield = NULL,
  projection = sp::CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs"))
}
\arguments{
\item{shape1}{An Spatial Points/Polygons Data Frame containing the geometry to add an attribute to}

\item{shape2}{An Spatial Points/Polygons Data Frame containing the geometry to add an attribute from}

\item{attributefield}{The name of the field in \code{shape2} as a string containing the values to add to \code{shape1}}

\item{newfield}{The name of the field in \code{shape1} as a string to add the values from \code{shape2$attributefield} to. If NULL, the field will use \code{attributefield}. Defaults to NULL.}

\item{projection}{An \code{sp::CRS()} argument. Defaults to \code{CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs")}}
}
\description{
This function will take a Spatial Points/Polygons Data Frame and add one attribute fields from a second Spatial Points/Polygons Data Frame
}
