% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Applying indicator benchmarks to a data frame of TerrADat data.}
\usage{
benchmark(benchmarks, tdat, evalstratumfield = "Evaluation.Stratum")
}
\arguments{
\item{benchmarks}{A data frame containing the benchmark information, imported from .XLSX with \code{read.benchmarks()}.}

\item{tdat}{A data frame imported from TerrADat's terrestrial or remote sensing indicators feature classes AND a field defining the evaluation stratum/group each plot should be considered part of, probably added using \code{attribute.shapefile()}.}

\item{evalstratumfield}{A string of the name of the field in \code{tdat} that contains the evaluation stratum/group identities. Defaults to \code{"Evaluation.Stratum"}.}
}
\value{
A data frame with the variables \code{"PRIMARYKEY", "PLOTID", "MANAGEMENT.QUESTION", "EVALUATION.STRATUM", "INDICATOR", "VALUE", "EVALUATION.CATEGORY"}
}
\description{
This evaluates the indicator values in TerrADat against the supplied benchmarks. If a plot should be evaluated in multiple evaluation strata/groups, there should be one copy of the plot per stratum/group.
}
\examples{
benchmark()
}
