% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ci_delta}
\alias{ci_delta}
\title{Calculate (un)transformed confidence intervals
This will calculate confidence intervals centered around the mean based on standard deviation, variance, or standard error. If more than one of those statistics is provided, they are defaulted to in that order.}
\usage{
ci_delta(
  mean,
  stddev = NULL,
  variance = NULL,
  stderr = NULL,
  transform = "none",
  alpha = 0.05,
  verbose = FALSE
)
}
\description{
Calculate (un)transformed confidence intervals
This will calculate confidence intervals centered around the mean based on standard deviation, variance, or standard error. If more than one of those statistics is provided, they are defaulted to in that order.
}
