% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ci_mean}
\alias{ci_mean}
\title{Calculate the upper and lower bounds for a mean given and alpha value}
\usage{
ci_mean(mean, sd, n, alpha)
}
\arguments{
\item{mean}{Numeric value. The mean to compute bounds for.}

\item{sd}{Numeric value. The standard deviation of \code{mean}.}

\item{n}{Numeric value. The number of observations that were used to calculate \code{mean}.}

\item{alpha}{Numeric value. The alpha value to use to compute the upper and lower confidence bounds for \code{mean}. Defaults to \code{0.05}.}
}
\value{
A named list containing the upper and lower bounds for the mean for the given confidence.
}
\description{
Calculate the upper and lower bounds for a mean given and alpha value
}
