% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{clip.arcpy}
\alias{clip.arcpy}
\title{Clip a Spatial Polygon/Points Data Frame to a Spatial Polygons Data Frame with ArcPy}
\usage{
clip.arcpy(spdf, spdf.clip, temp.path = NULL,
  python.search.path = "C:/Python27")
}
\arguments{
\item{spdf}{A Spatial Point or Spatial Polygons Data Frame to be clipped}

\item{spdf.clip}{A Spatial Polygons Data Frame to clip \code{spdf} by}

\item{temp.path}{Optional character string. A pre-existing filepath to use as a temporary folder to write files to. Defaults to a temporary directory from \code{tempdir()}.}

\item{python.search.path}{Optional character string. The filepath to search for pythonw.exe in. Defaults to "C:/Python27".}
}
\value{
The remaining geometry and data in \code{spdf} after being restricted to only overlap with \code{spdf.clip}.
}
\description{
Clip a Spatial Polygon/Points Data Frame to a Spatial Polygons Data Frame with ArcPy
}
