% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{clip_rgeos}
\alias{clip_rgeos}
\title{Use intersect_rgeos() to clip spatial points or polygons data frames}
\usage{
clip_rgeos(x, clip_polygons, verbose = FALSE, scale = 1e+05)
}
\arguments{
\item{x}{Spatial points or polygons data frame. The geometry to restrict.}

\item{clip_polygons}{Spatial polygons data frame. The geometry to restrict \code{x} to.}

\item{scale}{Numeric. A value to be passed to \code{rgeos::setScale()} to use for the steps of this function. It will be returned to the previous value before the function returns its output. Defaults to \code{1e5} (precision to five decimal places).}
}
\description{
Spatially restrict spatial points or polygons data frames without losing any information (other than the excluded geometry).
}
