% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{combine.ddpoints}
\alias{combine.ddpoints}
\title{Combine Multiple Design Database Point Data Frames
Combines all the points data frames into a single data frame or if there's only one data frame in the list, returns that data frame. It restricts the results to only the fields for primary key, plot ID, date visited, point fate, and panel.}
\usage{
combine.ddpoints(dd.points)
}
\arguments{
\item{dd.points}{List of data frames. The \code{pts} list from the output of \code{read.dd}.}
}
\value{
A single SPDF or data frame with the fields PRIMARYKEY, PLOTID, DATE.VISITED, FINAL_DESIG, and PANEL
}
\description{
Combine Multiple Design Database Point Data Frames
Combines all the points data frames into a single data frame or if there's only one data frame in the list, returns that data frame. It restricts the results to only the fields for primary key, plot ID, date visited, point fate, and panel.
}
