% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{combine_points}
\alias{combine_points}
\title{Combine multiple spatial points data frames}
\usage{
combine_points(points, add_missing = FALSE, projection = NULL)
}
\arguments{
\item{points}{List of spatial points data frames. There must be at least one variable common to every set of points in the list.}

\item{add_missing}{Logical. If \code{TRUE} then the missing variables will be added to the points, but with \code{NA} because there are no matching values. This prevents the loss of variables. Defaults to \code{FALSE}.}

\item{projection}{Optional CRS object. Used as the projection for the output. If \code{NULL} then the projection from the points at index 1 of \code{points} will be used instead. Defaults to \code{NULL}.}
}
\value{
Spatial points data frame with only variables common to all input points.
}
\description{
Combines a list of spatial points data frames into a single spatial points data frame with only the variables common to all points.
}
