% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{combine_tdat_design}
\alias{combine_tdat_design}
\title{Combine spatial points data frames from TerrADat and sample design databases}
\usage{
combine_tdat_design(terradat_points, design_points, by = "key",
  add_coords = TRUE, projection = NULL)
}
\arguments{
\item{design_points}{Spatial points data frame or list of spatial points data frames. These must be points with the variables from AIM terrestrial sample design databases.}

\item{by}{Character string. The method to compare the TerrADat points and the design points to identify the design points that correspond to the TerrADat points. If \code{"key"} then it will be by the TerADat primary keys. If those are not populated in the design points, then use \code{"spatial"} to assume that any points within 100 m of the TerrADat points was sampled. Defaults to \code{"key"}.}

\item{projection}{Optional CRS object. Used as the projection for the output. If \code{NULL} then the projection from the points at index 1 of \code{tdat_points} will be used instead. Defaults to \code{NULL}.}

\item{tdat_points}{Spatial points data frame or list of spatial points data frames. These must be points with the variables from TerrADat.}
}
\value{
Spatial points data frame with only variables common to all input points.
}
\description{
Combines a list of spatial points data frames into a single spatial points data frame with only the variables common to all points. This also drops all the design points which correspond to TerrADat points because we want to use the actual sampled locations to determine weighting.
}
