% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{dd.split}
\alias{dd.split}
\title{Splitting apart sample design databases containing multiple designs}
\usage{
dd.split(dd.list)
}
\arguments{
\item{dd.list}{The output from \code{read.dd()}}
}
\value{
The list \code{dd.list} with one SPDF per design in each of the lists \code{"sf"}, \code{"pts"}, and \code{"strata"}, maintaining the order of the data in those lists.
}
\description{
When a sample design database contains more than one design, as indicated by multiple sample frames, this function will (after importing with \code{read.dd()}) split the SPDFs from every multiple-design database into one SPDF per design and insert them at the indices where the original SPDF was. The naming format for these new SPDFs is "geodatabase filename_sample frame ID".
}
