% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{dissolve}
\alias{dissolve}
\title{Dissolve polygons according to unique identities}
\usage{
dissolve(polygons, dissolve_var)
}
\arguments{
\item{polygons}{Spatial Polygons Data Frame. Must have at least the variable matching the value of \code{dissolve_var}.}

\item{dissolve_var}{Character string. Must match the name of the variable in \code{polygons@data} that contains the unique identifiers to dissolve along.}
}
\value{
Spatial polygons data frame containing only the variable named in \code{dissolve_var} and only one observation per value in that variable.
}
\description{
For a spatial polygons data frame, return a spatial polygons data frame that contains only one observation for each unique identifier associated with the combined geometry of all polygons belonging to that identifier.
}
