% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{dpart_gen}
\alias{dpart_gen}
\title{Generate density partitions}
\usage{
dpart_gen(
  frame,
  points,
  n_dpart = 2,
  min_points = 1,
  iteration_limit = 1,
  density_sample_spacing = 100,
  buffer_distance = 1.55 * density_sample_spacing,
  max_cells_subset = 5e+05,
  output_list = FALSE,
  accept_failure = FALSE,
  sub_frame = TRUE,
  verbose = FALSE
)
}
\description{
Generate density partitions
}
