% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{erase.arcpy}
\alias{erase.arcpy}
\title{Erase geometry from a Spatial Polygons/Points Data Frame using a Spatial Polygons Data Frame with ArcPy}
\usage{
erase.arcpy(spdf, spdf.erase, temp.path = NULL,
  python.search.path = "C:/Python27")
}
\arguments{
\item{spdf}{A Spatial Polygons or Points Data Frame to remove FROM.}

\item{spdf.erase}{A Spatial Polygons Data Frame to remove geometry from \code{spdf} WITH.}

\item{temp.path}{Optional character string. This must be the path to a folder that R has write permissions to so that a subfolder called arcpy_temp can be created and used for ArcPy erasure steps. Defaults to the output from a \code{tempdir()} call.}

\item{python.search.path}{Optional character string. The filepath for the folder containing \code{pythonw.exe}. Defaults to \code{"C:/Python27"}.}
}
\value{
The remaining geometry and data in \code{spdf} after any parts overlapping \code{spdf.erase} has been removed from it.
}
\description{
Spatial manipulations in R can get, for lack of a better term, squirrelly. This removes the geometry of one SPDF from another via a system call to \code{Python} using the library \code{ArcPy}.
}
