% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{erase.rgeos}
\alias{erase.rgeos}
\title{Erase geometry from a Spatial Polygons/Points Data Frame using a Spatial Polygons Data Frame with rgeos}
\usage{
erase.rgeos(spdf, spdf.erase, sliverdrop = TRUE, sliverwarn = TRUE,
  sliverthreshold = 0.01)
}
\arguments{
\item{spdf}{A Spatial Polygons Data Frame to remove FROM.}

\item{spdf.erase}{A Spatial Polygons Data Frame to remove geometry from \code{spdf} WITH.}

\item{sliverdrop}{Optional logical value. If \code{erase} is \code{"rgeos"} this will be passed to \code{rgeos::set_RGEOS_dropSlivers()} to temporarily set the environment during the erasure attempt. Defaults to \code{TRUE}.}

\item{sliverwarn}{Optional logical value. If \code{erase} is \code{"rgeos"} this will be passed to \code{rgeos::set_RGEOS_warnSlivers()} to temporarily set the environment during the erasure attempt. Defaults to \code{TRUE}.}

\item{sliverdrop}{Optional numeric value. If \code{erase} is \code{"rgeos"} this will be passed to \code{rgeos::set_RGEOS_polyThreshold()} to temporarily set the environment during the erasure attempt. Defaults to \code{0.01}.}
}
\value{
The remaining geometry and data in \code{spdf} after any parts overlapping \code{spdf.erase} has been removed from it.
}
\description{
Erase geometry from a Spatial Polygons/Points Data Frame using a Spatial Polygons Data Frame with rgeos
}
