% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{fate.lookup}
\alias{fate.lookup}
\title{Return the default point fate lookup table}
\usage{
fate.lookup()
}
\description{
The default lookup table for final point fate designations in sample design databases.
}
