% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.analysis.R
\name{filename.aim}
\alias{filename.aim}
\title{Generating standard-format filenames for output from AIM analyses}
\usage{
filename.aim(name, type, extension = NULL)
}
\arguments{
\item{name}{A string of the project name, e.g. "Idaho_SageGrouse"}

\item{type}{A string of the data type, e.g. "analysis" or "reportingunits"}

\item{extension}{A string of the file extension, e.g. "csv" or "shp". If this argument is not provided, no extension will be appended.}
}
\description{
Creates an appropriate filename using the name of the project, the current date in the format DDMMYY, the type of data, and the file extension, separated by underscores.
}
