% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{filter.date}
\alias{filter.date}
\title{Filtering data frames by dates}
\usage{
\method{filter}{date}(data, date.field, after = NULL, before = NULL,
  ignore.year = FALSE)
}
\arguments{
\item{data}{Data frame. So long as \code{[]} will work on it, it's fine so spatial data frames are valid.}

\item{date.field}{Character string. The name of the column/variable in the data frame containing the date values. The date values must be interpretable by \code{lubridate::as_date()}.}

\item{after}{Optional character string expressing a date in the format \code{YYYY-MM-DD} (or \code{MM-DD} if \code{ignore.year = TRUE}). The earliest date (inclusive) to return data from. Rows containing earlier dates will not be returned.}

\item{before}{Optional character string expressing a date in the format \code{YYYY-MM-DD} (or \code{MM-DD} if \code{ignore.year = TRUE}). The latest date (inclusive) to return data from. Rows containing later dates will not be returned.}

\item{ignore.year}{Logical. If \code{TRUE} then years will be ignored in comparisons, e.g. If \code{after = "1986-10-30"} or \code{after = "10-30"} then all rows containing dates from October 30 to December 31 would be returned, regardless of the year. Defaults to \code{FALSE}.}
}
\description{
Given a date field and a lower date, an upper date, or both, filter data.
}
