% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{indicator.lookup}
\alias{indicator.lookup}
\title{Return the default indicator lookup table}
\usage{
indicator.lookup()
}
\description{
The default lookup table for combining the BLM Benchmark Tool table with the data from TerrADat
}
