% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{intersect.arcpy}
\alias{intersect.arcpy}
\title{Find the intersection of two Spatial Polygon/Points Data Frames with ArcPy}
\usage{
intersect.arcpy(spdf, spdf.intersect, temp.path = NULL,
  python.search.path = "C:/Python27")
}
\arguments{
\item{spdf}{A Spatial Polygons Data Frame}

\item{spdf.intersect}{A Spatial Polygons Data Frame to intersect with \code{spdf}.}

\item{temp.path}{Optional character string. A pre-existing filepath to use as a temporary folder to write files to. Defaults to a temporary directory from \code{tempdir()}.}

\item{python.search.path}{Optional character string. The filepath to search for pythonw.exe in. Defaults to "C:/Python27".}
}
