% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{intersect}
\alias{intersect}
\title{Create a SpatialPolygonsDataFrame from the intersection of two SpatialPolygonsDataFrames}
\usage{
intersect(spdf1, spdf1.attributefieldname,
  spdf1.attributefieldname.output = NULL, spdf2, spdf2.attributefieldname,
  spdf2.attributefieldname.output = NULL, method = "gintersection",
  area.ha = T, area.sqkm = T,
  projection = CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs"))
}
\arguments{
\item{spdf1}{A SpatialPolygonsDataFrame to be intersected}

\item{spdf1.attributefieldname}{Name of the field in \code{spdf1} unique to the unit groups or units to take values from.}

\item{spdf1.attributefieldname.output}{Optional string for the name of the field in the output SPDF to duplicate values from spdf1.attributefieldname into. If unspecified, then \code{spdf1.attributefieldname} will be used. Defaults to \code{NULL}.}

\item{spdf2}{A SpatialPolygonsDataFrame to be intersected}

\item{spdf2.attributefieldname}{Name of the field in \code{spdf2} unique to the unit groups or units to take values from.}

\item{spdf2.attributefieldname.output}{Optional string for the name of the field in the output SPDF to duplicate values from spdf2.attributefieldname into. If unspecified, then \code{spdf2.attributefieldname} will be used. Defaults to \code{NULL}.}

\item{method}{A string specifying which function to use for the interscting step: either \code{rgeos::gIntersection()} or \code{raster::intersect()}. Valid options are \code{"gintersection"} and \code{"intersect"}. Defaults to \code{"gintersection"}.}

\item{area.ha}{Logical. If \code{T}, areas will be calculated and added in hectares. Default is \code{T}.}

\item{area.sqkm}{Logical. If \code{T}, areas will be calculated and added in square kilometers. Default is \code{T}.}

\item{projection}{An \code{sp::CRS()} argument. The final output will be reprojected using this. Defaults to \code{CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs")}}
}
\value{
A SpatialPolygonsDataFrame with the attributes inherited from \code{spdf1} and \code{spdf2}, areas as appropriate, and a unique identifier.
}
\description{
Basically a wrapping of \code{rgeos::gIntersection()} and \code{raster::intersect()} with additional opportunity to call area.add() and automatically added unique identifiers for each resultant polygon
}
\examples{
intersect()
}
