% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{intersect_rgeos}
\alias{intersect_rgeos}
\title{Creating the Intersection of Two Spatial Polygons Data Frames}
\usage{
intersect_rgeos(
  polygons1,
  polygons2,
  keep_outer = FALSE,
  projection = NULL,
  makevalid = TRUE,
  force = FALSE,
  verbose = FALSE,
  scale = 1e+05
)
}
\arguments{
\item{polygons1}{Spatial polygons data frame. One of the set of polygons to intersect. If \code{crs} is \code{NULL} the projection from \code{polygons1@proj4string} will be used for the output.}

\item{polygons2}{Spatial polygons data frame. One of the set of polygons to intersect.}

\item{keep_outer}{Logical. If \code{TRUE} then the non-intersecting portions of \code{polygons1} and \code{polygons2} will be kept as well.}

\item{projection}{Optional \code{sp::CRS()} call. Used to reproject both \code{polygons1} and \code{polygons2} using \code{sp::spTransform()} If \code{NULL} then the projection from \code{polygons1@proj4string} will be used. Defaults to \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE} then additional informative messages will be produced as the function runs. Defaults to \code{FALSE}.}

\item{scale}{Numeric. A value to be passed to \code{rgeos::setScale()} to use for the steps of this function. It will be returned to the previous value before the function returns its output. Defaults to \code{1e5} (precision to five decimal places).}
}
\value{
A spatial polygons data frame consisting of the unique intersections of \code{polygons1} and \code{polygons2} with all the variables from both in the projection of either \code{projection} or \code{polygons1}. If there were variable names that occurred in both, those from \code{polygons2} will have a "2" appended.
}
\description{
Produce a spatial polygons data frame of the intersection of two spatial polygons data frames, inheriting all the variables from each. In the case of duplicate variables between the two, those in \code{polygons2@data} will have a "2" appended to differentiate them. This uses \code{rgeos::ginstersect} and assumes that the row names in both \code{polygons1} and \code{polygons2} DO NOT HAVE SPACES.
}
