% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{mulitplot_fate_distribution}
\alias{mulitplot_fate_distribution}
\title{Plotting Fate Distributions For Multiple Reporting Units
A wrapper for \code{plot_fate_distribution()} to apply that splits the input \code{point.weights} by the variable REPORTING.UNIT and returns a list}
\usage{
mulitplot_fate_distribution(point.weights, fates = NULL)
}
\arguments{
\item{point.weights}{Data frame. The point weights output from \code{weight()}.}

\item{fates}{Optional data frame. Defaults to the output from \code{fate.lookup()} which matches the plot fates found in sample design databases.}
}
\value{
A list of \code{ggplot} plots of the fate distributions of the reporting units as histograms.
}
\description{
Plotting Fate Distributions For Multiple Reporting Units
A wrapper for \code{plot_fate_distribution()} to apply that splits the input \code{point.weights} by the variable REPORTING.UNIT and returns a list
}
