% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{multiplot_indicator_distribution}
\alias{multiplot_indicator_distribution}
\title{Plotting Indicator Category Distributions For Multiple Indicators/Benchmarks/Reporting Units
A wrapper for plot_indicator_distribution to apply it across multiple combinations of values in Type, Subpopulation, MANAGEMENT.QUESTION, and Indicator simulataneously.}
\usage{
multiplot_indicator_distribution(analysis.df, benchmarks, indicator.lut = NULL)
}
\arguments{
\item{analysis.df}{Data frame. The output from \code{spsurvey::cat.analysis()} sliced to a single combination of the values in Type, Subpopulation, MANAGEMENT.QUESTION, and Indicator.}

\item{benchmarks}{Data frame. The output from \code{read.benchmarks()}.}

\item{indicator.lut}{Optional data frame. Defaults to the output from \code{indicator.lookup()} which matches the indicators found in TerrADat.}
}
\description{
Plotting Indicator Category Distributions For Multiple Indicators/Benchmarks/Reporting Units
A wrapper for plot_indicator_distribution to apply it across multiple combinations of values in Type, Subpopulation, MANAGEMENT.QUESTION, and Indicator simulataneously.
}
