% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{plot.date.distribution}
\alias{plot.date.distribution}
\title{Plotting Date Distributions}
\usage{
\method{plot}{date.distribution}(points.benchmarked)
}
\arguments{
\item{points.benchmarked}{Data frame. The output from \code{benchmark()}.}
}
\value{
A \code{ggplot} plot of the date distribution as a histogram.
}
\description{
Plotting Date Distributions
}
