% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{plot.indicator.distribution}
\alias{plot.indicator.distribution}
\title{Plotting Indicator Category Distributions}
\usage{
plot.indicator.distribution(analysis.df, benchmarks,
  indicator.lut = NULL, indicator.variable = "INDICATOR.TDAT",
  indicator.name = "INDICATOR.NAME")
}
\arguments{
\item{analysis.df}{Data frame. The output from \code{spsurvey::cat.analysis()} sliced to a single combination of the values in Type, Subpopulation, MANAGEMENT.QUESTION, and Indicator.}

\item{benchmarks}{Data frame. The output from \code{read.benchmarks()}.}

\item{indicator.lut}{Optional data frame. Defaults to the output from \code{indicator.lookup()} which matches the indicators found in TerrADat.}
}
\value{
A \code{ggplot} point plot with confidence interval bars.
}
\description{
Plotting Indicator Category Distributions
}
