% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{plot_fate_distribution}
\alias{plot_fate_distribution}
\title{Plotting Fate Distributions}
\usage{
plot_fate_distribution(point.weights, fates = NULL)
}
\arguments{
\item{point.weights}{Data frame. The point weights output from \code{weight()} with a common reporting unit.}

\item{fates}{Optional data frame. Defaults to the output from \code{fate.lookup()} which matches the plot fates found in sample design databases.}
}
\value{
A \code{ggplot} plot of the fate distribution as a histogram.
}
\description{
Plotting Fate Distributions
}
