% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{punchout}
\alias{punchout}
\title{Remove the geometry of one set of polygons from another}
\usage{
punchout(polygons, removal_polygons)
}
\arguments{
\item{polygons}{Spatial polygons data frame. The polygons to remove geometry from. Also the decider of the output projection.}

\item{removal_polygons}{Spatial polygons data frame. The geometry being removed from \code{polygons}.}
}
\description{
Using sf functions, find the difference in geometry between two spatial polygons data frames
}
