% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read.tdat}
\alias{read.tdat}
\title{Reading in an exported TerrADat geodatabase}
\usage{
read.tdat(tdat.path = NULL, tdat.name)
}
\arguments{
\item{tdat.path}{Character string. The folder path that contains the geodatabase specified in \code{tdat.name}. Defaults to the working directory.}

\item{tdat.name}{A string of the filename of the geodatabase to import from}
}
\description{
Reads in the terrestrial and remote sensing indicator feature classes from an ESRI geodatabase exported from the BLM NOC TerrADat. Returns a single SpatialPointsDataFrame containing the data from both.
}
