% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read.tracking}
\alias{read.tracking}
\title{Read in plot tracking Excel files}
\usage{
read.tracking(filename = "", path = "")
}
\arguments{
\item{filename}{A character vector of the filename (including extension) of the project tracking Excel file to import. If not using the \code{path} argument, the filename should include the entire filepath.}

\item{path}{Optional string specifying a the project tracking Excel file in \code{filename} to read in. This will be prepended to the value in \code{filename}. If the filepath is included in the string \code{filename}, do not provide this.}
}
\description{
Imports plot tracking Excel files.
}
