% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_benchmarks}
\alias{read_benchmarks}
\title{Reading in the benchmarks from the Data Explorer}
\usage{
read_benchmarks(filename = "", filepath = NULL,
  sheet_name = "Monitoring Objectives",
  eval_strings = list(c("Lower.Limit", "LL.Relation", "x"), c("x",
  "UL.Relation", "Upper.Limit"), c("x", "Proportion.Relation",
  "Required.Proportion")))
}
\arguments{
\item{filename}{Character string. The filename, including filetype extension, of the .XLSX, .CSV, .XLSM, or .XLS containing the benchmarks. Expects to find columns with headers matching "Management Question", "Benchmark Source", "Benchmark Group", "Reporting Unit", "Lower Limit", "LL Relation", "Indicator", "UL Relation", "Upper Limit", "Unit", "Condition Category", "Proportion Relation", and "Required Proportion".}

\item{filepath}{Optional character string. The filepath to the location where the file matching \code{filename} is stored. Used to locate the file by combining with the filename. Will be ignored if \code{NULL}. Defaults to \code{NULL}.}

\item{sheet_name}{Optional character string. The sheet name of the spreadsheet in the Excel workbook specified by \code{filename}. Only used if \code{filename} is an Excel workbook. Defaults to \code{"Monitoring Objectives"}.}

\item{eval_strings}{Optional list of character vectors. If \code{NULL}, nothing will be done. Otherwise, each character vector should contain one or more variable/column names from the benchmarks. The only other string allowed is \code{"x"} which can be used as a placeholder for indicator values. As an example, \code{list(c("lower.relationship", "lower.limit", "x"))} would add a column to the output called \code{"evalstring1"} that contains the results of \code{paste(benchmarks$lower.limit, benchmarks$lower.relationship, "x")}. Defaults to \code{list(c("Lower.Limit", "LL.Relation", "x"), c("x", "UL.Relation", "Upper.Limit"), c("x", "Required.Relation", "Required.Proportion"))}, appropriate for use with the AIM "Benchmark Tool".}
}
\value{
A data frame of the benchmarks from the specified file with fields containing evaluation strings to use in testing indicator values against the benchmarks.
}
\description{
Reading in the benchmarks from the Data Explorer
}
\examples{
read_benchmarks()
}
