% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_shapefile}
\alias{read_shapefile}
\title{Read in a shapefile}
\usage{
read_shapefile(filename, filepath = NULL, projection = NULL,
  stringsAsFactors = FALSE)
}
\arguments{
\item{filename}{Character string. The filename of the shapefile to read in. May be the full filepath to the shapefile, in which case use \code{filepath = NULL}. Does not need to include the file extension.}

\item{filepath}{Optional character string. The filepath to the location where the shapefile matching \code{filename} is stored. Used to locate the file by combining with the filename. Will be ignored if \code{NULL}. Defaults to \code{NULL}.}

\item{projection}{Optional CRS object. The projection to return the shapefile in as a CRS object, e.g. \code{sp::CRS("+proj=aea")}. Will be ignored if \code{NULL} or if it matches the shapefile's existing projection. Defaults to \code{NULL}.}

\item{stringsAsFactors}{Logical. To be passed to \code{rgdal::readOGR(stringsAsFactors)}. Defaults to \code{FALSE}.}
}
\value{
Spatial data frame.
}
\description{
Using \code{rgdal::readOGR()}, read in a shapefile.
}
