% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_shapefiles}
\alias{read_shapefiles}
\title{Read in multiple shapefiles}
\usage{
read_shapefiles(filenames, filepaths = NULL, projection = NULL,
  stringsAsFactors = FALSE)
}
\arguments{
\item{filenames}{Character string or vector of character strings. The filename(s) of the shapefile(s) to read in. May be the full filepath(s) to the shapefile(s), in which case use \code{filepath = NULL}. Does not need to include the file extension.}

\item{filepaths}{Optional character string or vector of character strings. The filepath(s) to the location(s) where the shapefile(s) matching \code{filename} is/are stored. Used to locate the file(s) by combining with the filename(s). Must be \code{NULL} (will be ignored), a single character vector (if only reading one shapefile or if all shapefiles are in the same location), or the same length as \code{filenames} (if the filenames are stored in different locations; the filenames and filepaths are paired by index in \code{filenames} and \code{filepaths}). Defaults to \code{NULL}.}

\item{projection}{Optional CRS object. The projection to return the shapefile(s) in as a CRS object, e.g. \code{sp::CRS("+proj=aea")}. Will be ignored if \code{NULL} or if it matches the shapefile's existing projection. Defaults to \code{NULL}.}

\item{stringsAsFactors}{Logical. To be passed to \code{rgdal::readOGR(stringsAsFactors)}. Defaults to \code{FALSE}.}
}
\value{
Spatial data frame.
}
\description{
Using \code{rgdal::readOGR()}, read in one or more shapefiles.
}
