% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{report}
\alias{report}
\title{Generating summary reports.}
\usage{
report(out.path, project.name, conf.level = 80, indicator.lut = NULL,
  benchmarks = NULL, analysis = NULL, cats.to.suppress = c("Not Meeting"),
  strata.stats, strata.weights, reporting.units.spdf = NULL,
  sample.frame.spdf, project.area.spdf = NULL, points.benchmarked = NULL,
  projection = sp::CRS("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"),
  extension = "html")
}
\arguments{
\item{out.path}{Character string. The folder path to write the output reports into.}

\item{project.name}{Character string. The name of the project to be used in the title of the report and the filenames it's written to.}

\item{conf.level}{Numeric. The confidence level expressed as an integer value, e.g. \code{80} if the confidence level is 80\%. Defaults to \code{80}.}

\item{indicator.lut}{Data frame. The lookup table between the indicator names in TerrADat and the human readable indicator names.}

\item{benchmarks}{Data frame. The output from \code{read.benchmarks()}.}

\item{analysis}{Data frame. The data frame \code{"analyses"} from the output of \code{analyze()}.}

\item{cats.to.suppress}{Character vector. One or more categories to suppress in tables. Defaults to \code{c("Not Meeting")}.}

\item{strata.stats}{Data frame. The data frame \code{strata.stats} from the output of \code{weight()}.}

\item{reporting.units.spdf}{Spatial polygons data frame. Used in plotting maps.}

\item{sample.frame.spdf}{Spatial polygons data frame. Used in plotting maps.}

\item{project.area.spdf}{Spatial polygons data frame. Used in plotting maps.}

\item{points.benchmarked}{Data frame. The output from \code{benchmark()}.}

\item{projection}{Optional \code{sp::CRS()} argument. Used to convert \code{points.benchmarked} into a spatial points data frame. Only specify if \code{points.benchmarked} has coordinates not from the same projection as TerrADat. Defaults to \code{sp::CRS("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0")}.}
}
\description{
Generate both a .PDF and .HTML report with \code{RMarkdown}.
}
