% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{restrict.tdat}
\alias{restrict.tdat}
\title{Restricting imported TerrADat data to the boundaries of the imported design databases}
\usage{
restrict.tdat(dd.raw, tdat.spdf)
}
\arguments{
\item{dd.raw}{The output from \code{read.dd()}. The function uses the spatial polygons data frames found in the \code{sf} list}

\item{tdat.spdf}{The output from \code{read.tdat()}.}
}
\value{
The contents of \code{read.tdat} that overlap spatially with the sample frames in \code{dd.raw}
}
\description{
Restricting imported TerrADat data to the boundaries of the imported design databases
}
