% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdd.reader.R
\name{sdd.reader}
\alias{sdd.reader}
\title{Importing Sample Design Databases for AIM Sample Designs}
\usage{
sdd.reader(src = "", sdd.src, func = "arcgisbinding", validate.keys = T,
  target.values = c("Target Sampled", "TS"), omitNAdesignations = F,
  projection = CRS("+proj=longlat +datum=NAD83 +no_defs +ellps=GRS80 +towgs84=0,0,0"))
}
\arguments{
\item{src}{Character string defining the filepath containing the sample design database[s]}

\item{sdd.src}{Character string or character vector containing the filenames of the geodatabases to import. Each filename should include the extension ".gdb"}

\item{func}{Character string. Defines whether to use the rgdal or arcgisbinding package to read in the geodatabases. Defaults to "arcgisbinding". Valid values are "arcgisbinding" and "readogr". This is not case sensitive}

\item{projection}{CRS string. Defaults to NAD83 CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs"). Is used to reproject all SPDFs in order to perform spatial manipulations}
}
\description{
This function imports one or more Sample Design Database[s] and returns a list of the named lists sf, pts, and strata. The named lists contain SpatialPoints/PolygonsDataFrames of the sample frame, points, and strata from each of the geodatabases. The SPDFs are named using the filename of the geodatabase source, so that each list has one SPDF named for each geodatabase imported and those names are identical between lists. If a Sample Design Database is missing any one of those features, a NULL value replaces the SPDF.
}
\examples{
weighter()
}
\keyword{weights}
