% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_categories.R
\name{sdd_clip}
\alias{sdd_clip}
\title{Write out a subset of polygon shapefiles from a sample design database}
\usage{
sdd_clip(
  reportingunit,
  reportingunit_path = NULL,
  sdd_path,
  lmf_strata = NULL,
  lmf_path = NULL,
  pointvisits = NULL,
  pointvisits_path = NULL,
  output_path,
  verbose = FALSE
)
}
\arguments{
\item{reportingunit}{An sf polygons object, Spatial Polygons Data Frame, or character string. If a character string, it must be the name of a polygon feature class found in \code{reportingunit_path}. Regardless, the feature class must contain exactly one feature representing a single reporting unit.}

\item{reportingunit_path}{Optional character string. The filepath to the folder or geodatabase containing the feature class specified with the character string \code{reportingunit}. Not necessary if \code{reportingunit} is either an sf polygon object or Spatial Polygons Data Frame. Defaults to \code{NULL}.}

\item{sdd_path}{Character string. The full filepath to the sample design database (SDD), including .GDB file extension. The SDD must follow the standard schema, including the feature classes: TerrestrialPointEvaluation, TerrestrialSamplePoints, TerrestrialStrata, and TerrestrialSampleFrame.}

\item{lmf_strata}{Optional sf polygons object, Spatial Polygons Data Frame, or character string. If a character string, it must be the name of a polygon feature class found in \code{reportingunit_path}. Regardless, the feature class must contain exactly one feature representing a single reporting unit. Defaults to \code{NULL}.}

\item{lmf_path}{Optional character string. If the weight categories need to account for the Landscape Monitoring Framework (LMF) strata, this must be the full filepath, including file extension, to the geodatabase containing the strata feature class with the LMF strata.}

\item{pointvisits}{Optional sf points object, Spatial Points Data Frame, or character string. If a character string, it must be the name of a point feature class found in \code{pointvisits_path}. Regardless, the feature class must contain the points to be considered and in the same format as the feature class TerrestrailPointEvaluation. This would only be used if you have already restricted which point evaluations should be considered for this process; otherwise, leave \code{NULL} and use the full set of points in the sample design database defined by \code{sdd_path}. Defaults to \code{NULL}.}

\item{pointvisits_path}{Optional character string. If \code{pointvisits} is a character string naming a feature class then this must be the path to the folder or geodatabase containing that feature class. Defaults to \code{NULL}.}

\item{output_path}{Character string. The filepath to the folder into which the resulting feature classes will be written.}

\item{verbose}{Logical. If \code{TRUE} then diagnostic messages will be returned throughout processing. Defaults to \code{FALSE}.}
}
\value{
Nothing. Writes out polygon shapefiles to a specified location.
}
\description{
This will take a reporting unit polygon and produce a set of polygons ready for intersection to generate weight categories using an appropriately formatted sample design database and, optionally, Landscape Management Framework (LMF) stratification polygons. Please note that all input polygons MUST be dissolved by their unique identifiers before use.
}
