% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{sf_intersect}
\alias{sf_intersect}
\title{Creating the Intersection of Two Spatial Polygons Data Frames using sf functions}
\usage{
sf_intersect(polygons1, polygons2, polygons1_attribute,
  polygons2_attribute, polygons1_attribute_output = NULL,
  polygons2_attribute_output = NULL, projection = NULL)
}
\arguments{
\item{polygons1}{Spatial polygons data frame. One of the two sets of geometry to intersect. \code{polygons1@data} must contain a variable with the name given as \code{polygons1_attribute}.}

\item{polygons2}{Spatial polygons data frame. One of the two sets of geometry to intersect. \code{polygons2@data} must contain a variable with the name given as \code{polygons2_attribute}.}

\item{polygons1_attribute}{Character string. The name of the variable in \code{polygons1@data} for the output to inherit values from.}

\item{polygons2_attribute}{Character string. The name of the variable in \code{polygons2@data} for the output to inherit values from.}

\item{polygons1_attribute_output}{Optional character string. The name of the variable in the output for the values inherited from \code{polygons1}. If \code{NULL} then the name from \code{polygons1_attribute} will be used. Defaults to \code{NULL}}

\item{polygons2_attribute_output}{Optional character string. The name of the variable in the output for the values inherited from \code{polygons2}. If \code{NULL} then the name from \code{polygons2_attribute} will be used. Defaults to \code{NULL}}

\item{projection}{Optional CRS object. Used as the projection for both \code{polygons1} and \code{polygons2}. If \code{NULL} then the \code{polygons@proj4string} will be used instead. Defaults to \code{NULL}.}
}
\description{
This uses \code{sf::st_intersection} to find ONLY the intersecting geometry of two SPDFs and return a single attribute variable from each.
}
