% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{table.reportingunits}
\alias{table.reportingunits}
\title{Creating tables of data}
\usage{
table.reportingunits(benchmarks, analysis.df, indicator.lut = NULL,
  output = "list")
}
\arguments{
\item{benchmarks}{Data frame. The output from \code{read.benchmarks()}.}

\item{analysis.df}{Data frame. The output from \code{spsurvey::cat.analysis()}}

\item{indicator.lut}{Optional data frame. Defaults to the output from \code{indicator.lookup()} which matches the indicators found in TerrADat.}

\item{output}{Character string. Determines the output type. The options are \code{"list"} which returns a list of data frames, \code{"dataframe"} which returns the same results, but as a single data frame, and \code{"xtables"} which returns a list of formatted tables using \code{xtable::xtable()}. Defaults to \code{"list"}}
}
\value{
A dataframe or a list of dataframes/xtables. Defaults to a list of dataframes.
}
\description{
Creating tables of data
}
