% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{thiessen_polygons_gen_fixed}
\alias{thiessen_polygons_gen_fixed}
\title{Create Thiessen/Voronoi polygons from a set of points and bounding polygons}
\usage{
thiessen_polygons_gen_fixed(
  centroids,
  frame,
  envelope = NULL,
  use_albers = TRUE
)
}
\arguments{
\item{centroids}{An sf points object. These points are used as centroids for the Thiessen/Voronoi polygons.}

\item{frame}{An sf polygon or multipolygon object. This is the clipping boundary which will be applied to the otherise "infinite" Thiessen/Voronoi polygons.}

\item{envelope}{An sfc polygon object. This will be the outer envelope for the Thiessen polygons before they're clipped to \code{frame}. This will only be applied if it's larger than the default envelope in \code{sf::st_voronoi()}. If \code{NULL} then the default envelope will be used. Defaults to \code{NULL}.}

\item{use_albers}{Logical. If \code{TRUE} then \code{centroids} and \code{frame} will be reprojected into Albers Equal Area (AEA) and the output will be in AEA. If \code{FALSE} then \code{frame} will be reprojected to match the coordinate reference ssytem (CRS) of \code{centroids} and the output will be in that CRS. CRSs using decimal degrees will throw errors or warnings. Defaults to \code{TRUE}.}
}
\value{
An sf object composed of polygon or multipolygon geometry
}
\description{
Generate Thiessen/Voronoi polygons for a set of points and clip the results using a set of polygons
}
