% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{tpoly_gen}
\alias{tpoly_gen}
\title{Create Thiessen/Voronoi polygons from a set of points and bounding polygons}
\usage{
tpoly_gen(
  frame,
  n_tpolys = 3,
  n_tpolys_min = n_tpolys,
  points = NULL,
  polygon_id_var = NULL,
  min_points = 1,
  n_tpoly_solutions = 1,
  accept_failure = FALSE,
  sub_frame = TRUE,
  envelope = NULL,
  seed_number = NULL,
  seed_increment = 1e+05,
  iteration_limit = 500,
  keep_crs = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{frame}{An sf polygon or multipolygon object. This is the clipping boundary which will be applied to the otherise "infinite" Thiessen/Voronoi polygons.}

\item{points}{Optional sf point object. If provided, then the Thiessen polygons will be redrawn with new random seeds until each contains at least \code{points_min} of these points. Defaults to \code{NULL}.}

\item{envelope}{An sfc polygon object. This will be the outer envelope for the Thiessen polygons before they're clipped to \code{frame}. This will only be applied if it's larger than the default envelope in \code{sf::st_voronoi()}. If \code{NULL} then the default envelope will be used. Defaults to \code{NULL}.}

\item{seed_number}{Optional numeric value. The seed number to use for generating the polygon centroids. A random seed will be used if this is \code{NULL}. Defaults to \code{NULL}.}

\item{seed_increment}{Optional numeric value. If attempting to produce polygons with \code{points_min} points from \code{points} in each polygon, this is the step to increment \code{seed_number} by on each attempt. Defaults to \code{100000}.}

\item{verbose}{Logical. If \code{TRUE} then the function will return diagnostic messages as it runs. Defaults to \code{FALSE}.)}

\item{n_polygons}{Numeric value. The number of Thiessen polygons to draw within the frame.}

\item{points_min}{Optional numeric value. If \code{points} is not \code{NULL} then this is the minimum number of points that each Thiessen polygon will contain. Defaults to \code{2}.}

\item{use_albers}{Logical. If \code{TRUE} then \code{centroids} and \code{frame} will be reprojected into Albers Equal Area (AEA) and the output will be in AEA. If \code{FALSE} then everything will be reprojected to match the coordinate reference system (CRS) of \code{frame} and the output will be in that CRS. CRSs using decimal degrees will throw errors or warnings. Defaults to \code{TRUE}.}
}
\value{
An sf object composed of polygon or multipolygon geometry
}
\description{
Generate Thiessen/Voronoi polygons for a set of points and clip the results using a set of polygons
}
