% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{validate.keys}
\alias{validate.keys}
\title{Confirming that design database points flagged as sampled have valid primary keys}
\usage{
validate.keys(dd.raw, target.values = c("Target Sampled", "TS"))
}
\description{
Confirming that design database points flagged as sampled have valid primary keys
}
