% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{weight.adjust}
\alias{weight.adjust}
\title{Adjusting Weights Calculated from AIM Sample Designs}
\usage{
weight.adjust(points, wgtcat.spdf, spdf.area.field = "AREA.HA.UNIT.SUM",
  spdf.wgtcat.field = "UNIQUE.IDENTIFIER",
  projection = sp::CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs"))
}
\arguments{
\item{points}{Data frame output from \code{weight()}, equivalent to \code{weight()[["point.weights"]]} or \code{weight()[[2]]}.}

\item{wgtcat.spdf}{SpatialPolygonsDataFrame describing the weight categories for adjusting the weights. Use the output from \code{intersect()}}

\item{spdf.area.field}{Character string defining the field name in \code{wgtcat@data} that contains the areas for the weight categories. Defaults to \code{"AREA.HA.UNIT.SUM"}.}

\item{spdf.wgtcat.field}{Character string defining the field name in \code{wgtcat@data} that contains the unique identification for the weight categories. Defaults to \code{"UNIQUE.IDENTIFIER"}.}

\item{projection}{\code{sp::CRS()} argument. Defaults to NAD83 with \code{sp::CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs")}. Is used to reproject all SPDFs in order to perform spatial manipulations}
}
\description{
This function takes the point weights data frame output from the function \code{weight()} and a SpatialPolygonsDataFrame defining the weight categories. Returns the data frame supplied as points with the new column \code{ADJWGT} containing the adjusted weights.
}
\examples{
weight.adjuster()
}
\keyword{weights}
