% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.adjuster.R
\name{weight.adjuster}
\alias{weight.adjuster}
\title{Adjusting Weights Calculated from AIM Sample Designs}
\usage{
weight.adjuster(points, wgtcat.spdf, spdf.area.field = "AREA.HA.UNIT.SUM",
  spdf.wgtcat.field = "UNIQUE.IDENTIFIER",
  projection = CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs"))
}
\arguments{
\item{points}{Data frame output from weighter(), equivalent to weighter()[["point.weights"]] or weighter()[[2]].}

\item{wgtcat.spdf}{SpatialPolygonsDataFrame describing the weight categories for adjusting the weights. Use the output from intersector()}

\item{spdf.area.field}{Character string defining the field name in wgtcat@data that contains the areas for the weight categories. Defaults to "AREA.HA.UNIT.SUM"}

\item{spdf.wgtcat.field}{Character string defining the field name in wgtcat@data that contains the unique identification for the weight categories. Defaults to "UNIQUE.IDENTIFIER"}

\item{projection}{CRS string. Defaults to NAD83 CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs"). Is used to reproject all SPDFs in order to perform spatial manipulations}
}
\description{
This function takes the point weights data frame output from the function weighter() and a SpatialPolygonsDataFrame defining the weight categories. Returns the data frame supplied as points with the new column ADJWGT containing the adjusted weights.
}
\examples{
weight.adjuster()
}
\keyword{weights}
