% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{weight_gen}
\alias{weight_gen}
\title{Generate weights using density partitions and/or Thiessen polygons}
\usage{
weight_gen(
  frame,
  points,
  n_dpart = 1,
  n_tpolys = 1,
  n_tpolys_min = n_tpolys,
  min_dpart_points = 1,
  min_tpoly_points = 1,
  n_tpoly_solutions = 1,
  seed_number = NULL,
  tpoly_iteration_limit = 1000,
  accept_failure = FALSE,
  sub_frame = TRUE,
  dpart_iteration_limit = 1,
  density_sample_spacing = 100,
  buffer_distance = density_sample_spacing * 0.55,
  max_cells_subset = 5e+05,
  return_polygons = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{frame}{SF polygon object. The spatial extent to create weights for. This will be subdivided into the weighting polygons according to the other parameters. Any points provided in \code{points} which fall outside this area will not be considered.}

\item{points}{SF points object. The points to generate weights for. Any points which do not fall within \code{frame} will be ignored.}

\item{n_dpart}{Numeric. The number of density partitions to divide the frame into. The specific sampling density ranges assigned to each partition is automatically determined. A value of \code{1} is equivalent to not partitioning by density. Default to \code{1}.}

\item{n_tpolys}{Numeric. The number of random Thiessen polygons to generate within the frame (if \code{n_dpart} is \code{1}) or within *each* density partition (if \code{n_dpart} is greater than \code{1}). Default to \code{1}.}

\item{n_tpolys_min}{Numeric. The minimum number of Thiessen polygons to accept within the frame (or each density partition). This is checked only if \code{accept_failure} is \code{FALSE}. Defaults to the value of \code{n_tpolys}.}

\item{min_dpart_points}{Numeric. The minimum number of points which must be found in each density partition. If not all density partitions meet this requirement, the function will return an error. Defaults to \code{1}.}

\item{min_tpoly_points}{Numeric. The minimum number of points which must be found in each Thiessen polygon. If not all density partitions meet this requirement, the function will return an error. Defaults to \code{1}.}

\item{n_tpoly_solutions}{Numeric. The number of unique, random Thiessen polygon solutions to find. Defaults to \code{1}.}

\item{seed_number}{Optional numeric. If this is not \code{NULL} it will be used to set seeds for all function calls which use randomness so that Thiessen polygon outputs are reproducible. Defaults to \code{NULL}.}

\item{tpoly_iteration_limit}{Numeric. The number of attempts to iterate on a Thiessen polygon seed number to find a valid solution (defined by \code{min_tpoly_points}) before abandoning the attempt. Higher values may be necessary for unevenly-distributed, low-density points. Defaults to \code{1000}.}

\item{accept_failure}{Logical. This determines if a failure to find a valid set of density partitions or Thiessen polygons will produce an error or be accepted. The intended use cases are those where the distribution of points makes the validity of any given set of random Theissen polygons particularly unlikely and it is more advantageous to find fewer Thiessen polygon solutions than to increase the value of \code{tpoly_iteration_limit} due to the increased computation time. Defaults to \code{FALSE}.}

\item{sub_frame}{Logical. If \code{accept_failure} is \code{TRUE}, this determines if the original \code{frame} be used when valid density partitions or Thiessen polygons were not produced. Defaults to \code{TRUE}.}

\item{dpart_iteration_limit}{Numeric. The number of times to increase the *generated* density partition count and consolidate the lowest-density partitions. This only happens when the low-density partitions contain fewer than \code{min_dpart_points} points and does not affect the number of *returned* density partitions. Defaults to \code{1}.}

\item{density_sample_spacing}{Numeric. The resolution of the density raster used to find partitions in meters. Higher values are coarser but computationally cheaper and faster. Defaults to \code{100}.}

\item{buffer_distance}{Numeric. The distance to buffer \code{frame} by when finding density to ensure that the resulting density partitions fully cover the original frame despite being raster-derived. Defaults to \code{density_sample_spacing * 0.55} to avoid parts of \code{frame} not being covered by cells.}

\item{max_cells_subset}{Numeric. If the number of cells in the density raster exceeds this value, the density partition break calculations will only use this many density values. This can dramatically increase calculation speeds. See \code{\link[BAMMtools:getJenksBreaks]{getJenksBreaks()}} documentation for additional details. Defaults to \code{500000}.}

\item{return_polygons}{Logical. If \code{TRUE} then the weighting polygons will be returned along with the weights. Defaults to \code{FALSE}.}

\item{verbose}{Logical. If \code{TRUE} then the function will generate additional messages as it executes. Defaults to \code{FALSE}.}
}
\value{
If \code{return_polygons} is \code{FALSE}, the \code{points} with their weights added. If \code{return_polygons} is \code{TRUE}, a named list containing the \code{points} with their weights added and the weighting polygons.
}
\description{
Produce sets of weights based on polygons created via density partitioning, random Thiessen polygons, or a combination of the two.
}
