% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzing.R
\name{weighted_variance}
\alias{weighted_variance}
\title{Calculate a weighted variance}
\usage{
weighted_variance(values, weights, na_remove = FALSE)
}
\arguments{
\item{values}{Numeric vector. The values to calculate the weighted variance for.}

\item{weights}{Numeric vector. The weights for the vector \code{values}. They must be in the same order as \code{values}.}

\item{na_remove}{Logical. If \code{TRUE} then any data with either a value or weight of \code{NA} will be removed before calculating. Defaults to \code{FALSE}.}
}
\description{
Calculate a weighted variance
}
