% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{write.analysis}
\alias{write.analysis}
\title{Writing out analysis outputs as a .csv}
\usage{
write.analysis(analysis.output, name, out.path)
}
\arguments{
\item{analysis.output}{Either a single output from \code{analyze()} or a list of multiple outputs, e.g. \code{c(analyze(dataset1), analyze(dataset2))}.}

\item{name}{A string of the project name, e.g. "Idaho_SageGrouse"}

\item{out.path}{A string specifying the output folder path}
}
\description{
Creates a \code{.csv} file composed of the output from one or more executions of \code{analyze()}. If combining multiple analyze() outputs, do so as a list with \code{c(analyze(dataset1), analyze(dataset2), ...)}.
}
