% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{write.benchmarkshp}
\alias{write.benchmarkshp}
\title{Writing out shapefiles of benchmarked points}
\usage{
write.benchmarkshp(points.benchmarked, tdat, out.path, name)
}
\arguments{
\item{points.benchmarked}{Data frame output from \code{benchmark()}.}

\item{tdat}{A Spatial Points Data Frame containing the relevant geometry and fields called \code{PLOTID} and \code{PRIMARYKEY}.}

\item{out.path}{A string specifying the output folder path.}

\item{name}{A string of the project name for use in the filename, e.g. "Idaho_SageGrouse".}
}
\description{
Writes out a point shapefile for each management question of points which were classified using the benchmarks for that question. Each
}
