% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.analysis.R
\name{write.shapefile}
\alias{write.shapefile}
\title{Writing out SPDFs as shapefiles}
\usage{
write.shapefile(spdf, dd = F, dd.list = "sf", union = T, name, type,
  outpath)
}
\arguments{
\item{spdf}{Either a single SPDF, a list of SPDFs, or the output from \code{read.dd()}. If multiple SPDFs are provided, they will be output as a single combined shapefile. All SPDFs must be of the same class. If not being combined with a \code{raster::union()} or if they are SpatialPointsDataFrames the \code{data} slots must have identical variables.}

\item{dd}{Logical. If the argument \code{spdf} is the output from \code{read.dd()}, provide \code{T}. Defaults to \code{F}.}

\item{dd.list}{A string specifying which of the three lists in dd to pull SPDFs from. Only used if \code{dd = T}. Valid values are \code{"sf"}, \code{"strata"}, and \code{"pts"}. Defaults to \code{"sf"}.}

\item{union}{Logical. If \code{T} and multiple SatialPolygonDataFrames are provided as \code{spdf} then \code{raster::union()} will be applied.}

\item{name}{A string of the project name for use in the filename, e.g. "Idaho_SageGrouse".}

\item{type}{A string of the type of output for use in the filename, e.g. "reportingunits" or "sampleframes".}

\item{out.path}{A string specifying the output folder path.}
}
\description{
Write our an ESRI shapefile from one or more SPDFs. If multiple SPDFs are being written out, they will be combined, optionally with \code{raster::union()} in the case of SpatialPolygonsDataFrame.
}
