% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzing.R
\name{analyze}
\alias{analyze}
\title{Categorical Analysis of AIM Data}
\usage{
analyze(benchmarked_points, point_weights, id_var,
  indicator_var = "indicator", value_var, weight_var = "wgt",
  x_var = "x", y_var = "y", reporting_var, split_var = NULL,
  conf = 80)
}
\arguments{
\item{benchmarked_points}{Data frame. This should be the output from \code{apply_benchmarks()} or an equivalent and must contain the variables specified in the arguments \code{id_var}, \code{value_var}, \code{indicator_var}, \code{x_var}, and \code{x_var}. If \code{split_var} is not \code{NULL} then that variable must also be in the data frame.}

\item{point_weights}{Data frame. This should be the \code{"point_weights"} data frame output from \code{weight()} or an equivalent. It must contain the variables specified in the arguments \code{weights.joinfield}, \code{weights.weightfield}, and \code{weights.reportingfield}.}

\item{id_var}{Character string. The name of the variable in \code{benchmarked_points} and \code{point_weights} that contains the values used to join them. The relationship of \code{point_weights}:\code{benchmarked_points} must be 1:1 or 1:Many.}

\item{indicator_var}{Character string. The name of the variable in \code{benchmarked_points} that contains the indicators for which there are categorical assignments in \code{benchmarked_points$value_var}.}

\item{value_var}{Character string. The name of the variable in \code{benchmarked_points} that contains the categorical value assignments for the indicators appearing in \code{benchmarked_points$indicator_var}.}

\item{x_var}{Character string. The name of the variable in \code{benchmarked_points} that contains the numeric values corresponding to the x coordinates (longitude) of the plots/observations.}

\item{y_var}{Character string. The name of the variable in \code{benchmarked_points} that contains the numeric values corresponding to the y coordinates (latitude) of the plots/observations.}

\item{reporting_var}{Character string. The name of the variable in \code{point_weights} that contains the values corresponding to the reporting unit (subpopulation) that the observations belong to.}

\item{split_var}{Optional character string or character vector. The name(s) of variable(s) that analyses should be split along. This must be done if a point had more than one benchmark for an indicator, e.g. splitting along the management objective variable if two different objects had distinct benchmarks for the same indicator that covered the same plot(s).}

\item{conf}{Numeric. The confidence level in percent. Defaults to \code{80}.}

\item{wgt_var}{Character string. The name of the variable in \code{point_weights} that contains the numeric weight values.}
}
\value{
A data frame with the output from \code{spsurvey::cat.analysis()}
}
\description{
This function takes the outputs from the functions \code{weight()} optionally passed through \code{weight.adjust()}, \code{benchmarker()}, and a TerrADat SpatialPointsDataFrame with a field for reporting units. It applies \code{spsurvey::cat.analysis()} and returns the results
}
\examples{
analyze()
}
\keyword{analysis}
