% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzing.R
\name{analyze}
\alias{analyze}
\title{Categorical Analysis of AIM Data}
\usage{
analyze(evaluated.points, point.weights,
  default.reportingunit = "No reporting unit", reportingunit.type = NA,
  adjustedweights = FALSE)
}
\arguments{
\item{evaluated.points}{Data frame output from \code{benchmark()}.}

\item{point.weights}{Data frame output from \code{weight()}, equivalent to \code{weight()[["point.weights"]]} or \code{weight()[[2]]}.}

\item{default.reportingunit}{A string to populate the reporting unit field with in the case that designs were not restricted by reporting unit. Defaults to \code{"No reporting unit"}}

\item{reportingunit.type}{A string to populate the reporting unit type field with. Depends on the reporting units used, e.g. it may be \code{"Watershed"} or \code{"Sage-grouse Habitat"}. Defaults to \code{NA}.}

\item{adjustedweights}{Logical. If \code{TRUE}, use the values in the field ADJWGT instead of the field WGT. Allows for quick comparison between the results of analysis with and without the weight adjustment. Defaults to \code{FALSE}.}
}
\value{
A data frame with the output from \code{spsurvey::cat.analysis()}
}
\description{
This function takes the outputs from the functions \code{weight()} optionally passed through \code{weight.adjust()}, \code{benchmarker()}, and a TerrADat SpatialPointsDataFrame with a field for reporting units. It applies \code{spsurvey::cat.analysis()} and returns the results
}
\examples{
analyze()
}
\keyword{analysis}
