% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzing.R
\name{analyze_con}
\alias{analyze_con}
\title{Estimation of weighted means of continuous data}
\usage{
analyze_con(
  data,
  weights,
  id_var,
  value_var,
  wgt_var,
  conf = 80,
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame. Continuous data (e.g., numeric) with the unique identifiers for each observation/row in the variable \code{id_var} and the value for each observation/row in \code{value_var}. Note that the unique identifiers are the link between \code{data} and \code{weights}}

\item{weights}{Data frame. This must contain the weighting information using the variables \code{id_var} with a unique identifier for each observation/row and \code{wgt_var} with the relative numeric weight of each observation/row.}

\item{id_var}{Character string. The name of the variable in \code{data} and \code{weights} that contains the unique identifiers for the observations. All values in \code{data$id_var} must appear in \code{weights$id_var}.}

\item{value_var}{Character string. The name of the variable in \code{data} that contains the values as character strings.}

\item{wgt_var}{Character string. The name of the variable in \code{weights} that contains the numeric weight values.}

\item{conf}{Numeric. The confidence level in percent. Defaults to \code{80}.}

\item{verbose}{Logical. If \code{TRUE} then the function will generate additional messages as it executes. Defaults to \code{FALSE}.}
}
\value{
A data frame containing the count of observations, weighted mean, and confidence intervals.
}
\description{
Given continuous data and the weights for the individual observations, calculate estimated mean and confidence intervals.
}
