% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzing.R
\name{analyze_weighted}
\alias{analyze_weighted}
\title{Weighted analyses}
\usage{
analyze_weighted(
  data,
  weights = NULL,
  id_vars,
  indicator_type,
  indicator_var = "indicator",
  value_var = "value",
  possible_categorical_values = NULL,
  conf = 80,
  combine = c("none", "delta", "bootstrap", "mean", "mean_weights"),
  bootstrap_replicates = 10000,
  bootstrap_type = "bca",
  verbose = FALSE
)
}
\description{
Weighted analyses
}
