% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{apply_benchmarks}
\alias{apply_benchmarks}
\title{Apply benchmarks to data}
\usage{
apply_benchmarks(
  data,
  idvars,
  benchmark_group_var = NULL,
  benchmark_group_lut = NULL,
  benchmarks,
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame or spatial * data frame. The data to be benchmarked. Must contain the variable(s) in \code{idvars} as unique identifiers. May be "wide" or "long" data. If wide, then each indicator name should be a variable name. If long, then the indicator names and values should be in variables named \code{"indicator_var"} and \code{"value"}.}

\item{idvars}{Character string or vecor of character strings. The names of the variable(s) in \code{data} which serve as unique identifiers.}

\item{benchmark_group_var}{Optional character string. The name of the variable that holds the benchmark group memberships. If \code{NULL} then it defaults to the variable in \code{benchmark_group_lut} that is not in \code{idvars}. Defaults to \code{NULL}.}

\item{benchmark_group_lut}{Optional data frame. The lookup table used to assign benchmark group membership to \code{data}. Must contain all of the variables in \code{idvar} and one additional variable with the benchmark group memberships.}

\item{benchmarks}{Data frame. The benchmarks in the format produced by \code{read.benchmarks()}. If the benchmark group membership variable in other data frames is not \code{"Benchmark.Group"} then it must be renamed here.}

\item{verbose}{Logical. If \code{TRUE} then the function will generate additional messages as it executes.}
}
\value{
A long data frame of each point/benchmarked indicator combination with the assigned condition category.
}
\description{
Using data that has unique identifier and indicator value variables, apply benchmarks. The benchmarks are applied to the data based on the benchmark group(s) that the data belong to, which may already be assigned in the data or may be added with a benchmark group lookup table.
}
