% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manipulation.R
\name{attribute.shapefile}
\alias{attribute.shapefile}
\title{Attributing a spatial data frame using another spatial data frame}
\usage{
attribute.shapefile(spdf1, spdf2, attributefield = NULL, newfield = NULL,
  projection = CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs"))
}
\arguments{
\item{spdf1}{A SpatialPoints/PolygonsDataFrame containing the geometry to add an attribute to}

\item{spdf2}{A SpatialPoints/PolygonsDataFrame containing the geometry to add an attribute from}

\item{attributefield}{The name of the field in \code{spdf2} as a string containing the values to add to \code{spdf1}}

\item{newfield}{The name of the field in \code{spdf1} as a string to add the values from \code{spdf2$attributefield} to. If NULL, the field will use \code{attributefield}. Defaults to NULL.}

\item{projection}{An \code{sp::CRS()} argument to apply in the event that \code{spdf1} and \code{spdf2} have different projections. Defaults to \code{CRS("+proj=longlat +ellps=GRS80 +datum=NAD83 +no_defs")}}
}
\value{
The original SPDF spdf1 with the new field containing the values inherited from spdf2.
}
\description{
This function will take a SpatialPoints/PolygonsDataFrame and add one attribute fields from a second SpatialPoints/PolygonsDataFrame
}
\examples{
attribute.shapefile()
}
