% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{check_polygons}
\alias{check_polygons}
\title{Confirm polygons contain points}
\usage{
check_polygons(
  polygons,
  points,
  minimum = 1,
  polygon_id_var = NULL,
  verbose = FALSE
)
}
\description{
Confirm polygons contain points
}
