% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzing.R
\name{ci_delta}
\alias{ci_delta}
\title{Calcualte (un)transformed CIs}
\usage{
ci_delta(
  mean,
  stddev = NULL,
  variance = NULL,
  transform = "none",
  alpha = 0.05
)
}
\description{
Calcualte (un)transformed CIs
}
