% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.benchmarks.R
\name{read.benchmarks}
\alias{read.benchmarks}
\title{Reading in the benchmarks from the Data Explorer}
\usage{
read.benchmarks(data.path = "", benchmarks.filename = "",
  sheet.name = "Monitoring Objectives", indicator.lut = NULL,
  indicator.lut.benchmarkfield = "indicator.name", convert.l2r = T)
}
\arguments{
\item{data.path}{A string specifying the path to the folder containing the .XLSX with the benchmarks}

\item{benchmarks.filename}{A string specifying the filename of the .XLSX containing the benchmarks}

\item{indicator.lut}{A data frame with the column \code{"indicator.name"} matching the values in the Data Explorer "Indicator" field and one called \code{"indicator.tdat"} with corresponding value for the indicators' names in TerrADat.}

\item{indicator.lut.benchmarkfield}{The name of the column in \code{indicator.lut} that matches the "Indicator column of Data Explorer. Defaults to \code{"indicator.name"}}
}
\value{
A data frame of the benchmarks from the Data Explorer with a field containing an evaluation string to use in testing indicator values against the benchmarks.
}
\description{
Reading in the benchmarks from the Data Explorer
}
\examples{
read.benchmarks()
}
