% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{report}
\alias{report}
\title{Generating summary reports.}
\usage{
report(out.path, project.name, indicator.lut = NULL, fates = NULL,
  benchmarks = NULL, analysis = NULL,
  cats.to.suppress = c("Not Meeting"), point.weights, strata.weights,
  reporting.units.spdf = NULL, sample.frame.spdf,
  project.area.spdf = NULL, points.benchmarked = NULL, dd.points,
  daterange.max = NULL, daterange.min = NULL,
  projection = sp::CRS("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0"),
  extension = "html")
}
\arguments{
\item{out.path}{Character string. The folder path to write the output reports into.}

\item{project.name}{Character string. The name of the project to be used in the title of the report and the filenames it's written to.}

\item{indicator.lut}{Data frame. The lookup table between the indicator names in TerrADat and the human readable indicator names.}

\item{benchmarks}{Data frame. The output from \code{read.benchmarks()}.}

\item{analysis}{Data frame. The data frame \code{"analyses"} from the output of \code{analyze()}.}

\item{cats.to.suppress}{Character vector. One or more categories to suppress in tables. Defaults to \code{c("Not Meeting")}.}

\item{point.weights}{Data frame. The data frame \code{point.weights} from the output of \code{weight()}.}

\item{strata.weights}{Data frame. The data frame \code{strata.weights} from the output of \code{weight()}.}

\item{reporting.units.spdf}{Spatial polygons data frame. Used in plotting maps. This MUST have a field named exactly "Type" containing the type of reporting unit (e.g. "Watershed" or "Study Area") and a field named exactly "Subpopulation" which contains the identity of the reporting unit[s] (e.g. the watersheds "Dickshooter Creek" and "Headwaters Deep Creek").}

\item{sample.frame.spdf}{Spatial polygons data frame. Used in plotting maps.}

\item{project.area.spdf}{Spatial polygons data frame. Used in plotting maps.}

\item{points.benchmarked}{Data frame. The output from \code{benchmark()}.}

\item{dd.points}{Spatial points data frame or list of spatial points data frames. This should be the \code{pts} list in the output from \code{read.dd()}.}

\item{daterange.max}{Optional character string. This must be interpretable by \code{lubridate::as_date()}, e.g. \code{"2016-04-20"}. Only sampling locations visited before this date will be considered. Currently only restricts to year.}

\item{daterange.min}{Optional character string. This must be interpretable by \code{lubridate::as_date()}, e.g. \code{"2016-04-20"}. Only sampling locations visited after this date will be considered. Currently only restricts to year.}

\item{projection}{Optional \code{sp::CRS()} argument. Used to convert \code{points.benchmarked} into a spatial points data frame. Only specify if \code{points.benchmarked} has coordinates not from the same projection as TerrADat. Defaults to \code{sp::CRS("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs +towgs84=0,0,0")}.}

\item{extension}{Character string. The file extension to write the report to. Defaults to \code{"html"}.}
}
\description{
Generate both a .PDF and .HTML report with \code{RMarkdown}.
}
