% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{restrict.tdat}
\alias{restrict.tdat}
\title{Restricting imported TerrADat data to the boundaries of the imported design databases}
\usage{
restrict.tdat(dd.raw, tdat.spdf)
}
\description{
Restricting imported TerrADat data to the boundaries of the imported design databases
}
